This should fix the tf-idf

+++ java/label/src/main/java/com/codeu/team34/label/JedisIndex.java
@@ -154,7 +154,7 @@ public class JedisIndex {
 		Set<String> docSet = jedis.smembers(setKey);
 		int numDocsContaining = docSet.size();
 		
-		tfidf = (0.5+0.5*(tf/docSize))*Math.log(numDocs/numDocsContaining);
+		tfidf = (0.5+0.5*((double)tf/(double)docSize))*Math.log((double)numDocs/(double)numDocsContaining);
 		
 		return tfidf;
 	}
